remove solution model and clean up
